Do not precompile the Criteria function
Storing the fun in the mnesia table is a little risky.  Therefore, we
change to a more conservative model that stores the expression as a
string and JITs the string at run time.  This is expected to be slower,
of course, but more stable.  We can optimize with caching the compiler
results later, if necessary.  The good news is that the compilation
of each criteria happens as part of a concurrent algorithm, which may
help to mitigate the overhead.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>